StarGobang/python/requirements.txt
# 协议验证：每行注释标注 PyPI 官方协议
# 【更新】适配 Python 3.12
torch>=2.2.0              # BSD-3-Clause
torchvision>=0.17.0       # BSD-3-Clause
numpy>=1.24.0             # BSD-3-Clause
onnx>=1.14.0              # MIT
onnxruntime>=1.16.0       # MIT
tqdm>=4.66.0              # MIT
scikit-learn>=1.3.0       # BSD-3-Clause (仅 arena 评估用)
pyyaml>=6.0               # MIT
# 【硬约束】禁止出现：gym, tensorflow, keras, opencv-python (GPL 风险)
